backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-socketio==5.11.1
aiohttp==3.9.3
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.0
joblib==1.3.2
pydantic==2.6.1
onnxmltools==1.13.0
groq==0.9.0
httpx==0.27.2
scipy>=1.12.0
statsmodels>=0.14.1
